backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
numpy==1.26.3
pillow==10.2.0
face_recognition==1.3.0
opencv-python==4.9.0.80
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.19.0
requests
pandas
openpyxl
